MTV-5700 | Move Disk.MaxLUN validation into ssh.go
Move validateDiskMaxLUN from maxlun.go into ssh.go alongside the
other XCOPY copy-offload preflight checks.

Ref: https://redhat.atlassian.net/browse/MTV-5700
Resolves: MTV-5700
